+++ piecrust/tests/eventer.rs
@@ -20,9 +20,10 @@ pub fn vm_center_events() -> Result<(), Error> {
 
     const EVENT_NUM: u32 = 5;
 
-    session.call(eventer_id, "emit_events", &EVENT_NUM)?;
+    let receipt =
+        session.call::<_, ()>(eventer_id, "emit_events", &EVENT_NUM)?;
 
-    let events = session.take_events();
+    let events = receipt.events;
     assert_eq!(events.len() as u32, EVENT_NUM);
 
     for i in 0..EVENT_NUM {